+++ xs/sources/lexer.c
@@ -415,6 +415,8 @@ XsToken xs_lexer_next(XsLexer *lexer)
     if(peek(lexer, 0) == '|')
       return one_or_two(lexer, start, '|', XS_TOKEN_PIPE, XS_TOKEN_LOGICAL_OR);
     return one_or_two(lexer, start, '=', XS_TOKEN_PIPE, XS_TOKEN_PIPE_ASSIGN);
+  case '^':
+    return one_or_two(lexer, start, '=', XS_TOKEN_CARET, XS_TOKEN_CARET_ASSIGN);
   case '@':
     return token(XS_TOKEN_AT, start, lexer->cursor);
   case '$':