+++ xs/sources/token.c
@@ -134,6 +134,8 @@ const char *xs_token_kind_name(XsTokenKind kind)
       "|",
       "||",
       "|=",
+      "^",
+      "^=",
       "@",
       "$",
       "as",